requirements.txt
absl-py==1.2.0
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
cycler==0.11.0
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1649586402495/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1658852325129/work
Flask==2.1.3
Flask-WTF==1.0.1
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpc-split_1658277749120/work
h5py @ file:///Users/runner/miniforge3/conda-bld/h5py_1637964045648/work
idna==3.3
importlib-metadata==4.12.0
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1657295113967/work
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1653754989223/work
itsdangerous==2.1.2
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1649067387646/work
Jinja2==3.1.2
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1658332720484/work
keras==2.9.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.1
llvmlite==0.39.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
numba==0.56.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1653325964689/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.4.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.2.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1656332401605/work
protobuf==3.19.4
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1653089402647/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1650904496387/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.1
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1656183620094/work
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
shap==0.41.0
six==1.15.0
slicer==0.0.7
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1655315839047/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1656938068018/work
tqdm==4.64.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1655411388954/work
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.2.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
